+++ src/Uno.UI/UI/Xaml/Controls/ListViewBase/ListViewBase.cs
@@ -684,221 +684,6 @@ namespace Windows.UI.Xaml.Controls
 			base.ContainerClearedForItem(item, itemContainer);
 		}
 
-		public event DragItemsStartingEventHandler DragItemsStarting;
-		public event TypedEventHandler<ListViewBase, DragItemsCompletedEventArgs> DragItemsCompleted;
-
-		public static DependencyProperty CanReorderItemsProperty { get; } = DependencyProperty.Register(
-			nameof(CanReorderItems),
-			typeof(bool),
-			typeof(ListViewBase),
-			new FrameworkPropertyMetadata(default(bool)));
-
-		public bool CanReorderItems
-		{
-			get => (bool)GetValue(CanReorderItemsProperty);
-			set => SetValue(CanReorderItemsProperty, value);
-		}
-
-		public static DependencyProperty CanDragItemsProperty { get; } = DependencyProperty.Register(
-			nameof(CanDragItems),
-			typeof(bool),
-			typeof(ListViewBase),
-			new FrameworkPropertyMetadata(default(bool)));
-
-		public bool CanDragItems
-		{
-			get => (bool)GetValue(CanDragItemsProperty);
-			set => SetValue(CanDragItemsProperty, value);
-		}
-
-		private static void OnItemContainerDragStarting(UIElement sender, DragStartingEventArgs innerArgs)
-		{
-			if (ItemsControlFromItemContainer(sender) is ListViewBase that && that.CanDragItems)
-			{
-				var items = that.SelectedItems.ToList();
-				var draggedItem = that.ItemFromContainer(sender);
-				if (draggedItem is { } && !items.Contains(draggedItem))
-				{
-					items.Add(draggedItem);
-				}
-
-				var args = new DragItemsStartingEventArgs(innerArgs, items);
-
-				that.DragItemsStarting?.Invoke(that, args);
-
-				// The application has the ability to add some items in the list, so make sure to freeze it only after event has been raised.
-				args.Data.SetData(DragItemsFormatId, args.Items.ToList());
-
-				// The ListView must have both CanReorderItems and AllowDrop flags set to allow re-ordering (UWP)
-				// We also do not allow re-ordering if we where not able to find the item (as it has to be hidden in the view) (Uno only)
-				if (that.CanReorderItems && that.AllowDrop && draggedItem is {})
-				{
-					args.Data.SetData(ReorderOwnerFormatId, that);
-					args.Data.SetData(ReorderItemFormatId, draggedItem);
-					args.Data.SetData(ReorderContainerFormatId, sender);
-
-					that.DragEnter += OnReorderUpdated;
-					that.DragOver += OnReorderUpdated;
-					that.DragLeave += OnReorderCompleted;
-					that.Drop += OnReorderCompleted;
-				}
-			}
-		}
-
-		private static void OnReorderUpdated(object sender, DragEventArgs dragEventArgs)
-		{
-			var that = sender as ListView;
-			var src = dragEventArgs.DataView.FindRawData(ReorderOwnerFormatId) as ListView;
-			var item = dragEventArgs.DataView.FindRawData(ReorderItemFormatId);
-			var container = dragEventArgs.DataView.FindRawData(ReorderContainerFormatId) as FrameworkElement; // TODO: This might have changed/been recycled if scrolled 
-			if (that is null || src is null || item is null || container is null || src != that)
-			{
-				dragEventArgs.Log().Warn("Invalid reorder event.");
-
-				return;
-			}
-
-			that.UpdateReordering(dragEventArgs.GetPosition(that), container, item);
-		}
-
-		private static void OnReorderCompleted(object sender, DragEventArgs dragEventArgs)
-		{
-			var that = sender as ListView;
-			var src = dragEventArgs.DataView.FindRawData(ReorderOwnerFormatId) as ListView;
-			var item = dragEventArgs.DataView.FindRawData(ReorderItemFormatId);
-			var container = dragEventArgs.DataView.FindRawData(ReorderContainerFormatId) as FrameworkElement; // TODO: This might have changed/been recycled if scrolled 
-			if (that is null || src is null || item is null || container is null || src != that)
-			{
-				dragEventArgs.Log().Warn("Invalid reorder event.");
-
-				return;
-			}
-
-			var updatedIndex = default(IndexPath?);
-			that.CompleteReordering(container, item, ref updatedIndex);
-
-			if (!that.IsGrouping
-				&& updatedIndex.HasValue
-				&& dragEventArgs.DataView.FindRawData(DragItemsFormatId) is List<object> movedItems)
-			{
-				if (that.UnwrapItemsSource() is {} unwrappedSource)
-				{
-					// The UWP ListView seems to automatically push back changes only if the ItemsSource inherits from ObservableCollection
-					var srcType = unwrappedSource.GetType();
-					if (srcType.IsGenericType
-						&& srcType.GetGenericTypeDefinition() == typeof(ObservableCollection<>)
-						&& srcType.GetMethod(nameof(ObservableCollection<object>.Move), new[] { typeof(int), typeof(int) }) is {} mv)
-					{
-						ProcessMove(
-							((ICollection)unwrappedSource).Count,
-							((ICollection)unwrappedSource).IndexOf,
-							(oldIndex, newIndex) => mv.Invoke(unwrappedSource, new object[] { oldIndex, newIndex }));
-					}
-				}
-				else // The ListView was created with items defined in XAML
-				{
-					var items = that.Items;
-					ProcessMove(
-						items.Count,
-						items.IndexOf,
-						(oldIndex, newIndex) =>
-						{
-							var item = items[oldIndex];
-							items.RemoveAt(oldIndex);
-							if (newIndex >= items.Count)
-							{
-								items.Add(item);
-							}
-							else
-							{
-								items.Insert(newIndex, item);
-							}
-						});
-				}
-
-				void ProcessMove(
-					int count,
-					Func<object, int> indexOf,
-					Action<int, int> mv)
-				{
-					var indexOfDraggedItem = indexOf(item);
-					if (indexOfDraggedItem < 0)
-					{
-						return;
-					}
-
-					int newIndex;
-					if (updatedIndex.Value.Row == int.MaxValue) 
-					{
-						// I.e. we are at the end, there is no items below
-						newIndex = count - 1;
-					}
-					else
-					{
-						newIndex = that.GetIndexFromIndexPath(updatedIndex.Value);
-						if (indexOfDraggedItem < newIndex)
-						{
-							// If we've moved items down, we have to take in consideration that the updatedIndex
-							// is already assuming that the item has been removed, so it's offsetted by 1.
-							newIndex--;
-						}
-					}
-
-					for (var i = 0; i < movedItems.Count; i++)
-					{
-						var movedItem = movedItems[i];
-						var oldIndex = indexOf(movedItem);
-
-						if (oldIndex < 0 || oldIndex == newIndex)
-						{
-							continue; // Item removed or already at the right place, nothing to do.
-						}
-
-						mv(oldIndex, newIndex);
-
-						if (oldIndex > newIndex)
-						{
-							newIndex++;
-						}
-					}
-				}
-			}
-		}
-
-		/// <summary>
-		/// Update reordering information
-		/// </summary>
-		/// <param name="location">The current location of the pointer use to reordering items, in the ListView coordinates space</param>
-		/// <param name="draggedContainer">The container that has been clicked to initiate the reordering (i.e. drag) operation (cf. remarks)</param>
-		/// <param name="draggedItem">The item that has been clicked to initiate the reordering (i.e. drag) operation (cf. remarks)</param>
-		/// <remarks>
-		/// If the SelectionMode is not None or Single, the draggedItem/Container might not be the single that is being reordered.
-		/// However, UWP hides in the ListView only the item that is being clicked by the user to initiate the reorder / drag operation.
-		/// </remarks>
-		partial void UpdateReordering(Point location, FrameworkElement draggedContainer, object draggedItem);
-		partial void CompleteReordering(FrameworkElement draggedContainer, object draggedItem, ref IndexPath? updatedIndex);
-
-		private static void OnItemContainerDragCompleted(UIElement sender, DropCompletedEventArgs innerArgs)
-		{
-			// Note: It's not the responsibility of the ListView to remove item from the source, not matter the AcceptedOperation.
-
-			if (ItemsControlFromItemContainer(sender) is ListViewBase that)
-			{
-				if (that.CanDragItems)
-				{
-					var items = innerArgs.Info.Data.FindRawData(DragItemsFormatId) as IReadOnlyList<object> ?? new List<object>(0);
-					var args = new DragItemsCompletedEventArgs(innerArgs, items);
-
-					that.DragItemsCompleted?.Invoke(that, args);
-				}
-
-				that.DragEnter -= OnReorderUpdated;
-				that.DragOver -= OnReorderUpdated;
-				that.DragLeave -= OnReorderCompleted;
-				that.Drop -= OnReorderCompleted;
-			}
-		}
-
 		/// <summary>
 		/// Apply the multi-selection state to the provided item
 		/// </summary>
